pyproject.toml
[project]
name = "second-brain"
version = "0.1.0"
description = "AI-powered journaling with smart prompts and automatic backlinks"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.15.1",
    "rich>=13.9.4",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.1",
    "azure-search-documents>=11.6.0b7",
    "openai>=1.57.2",
]

[project.scripts]
brain = "brain_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["brain_core", "brain_cli"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
